added a PieceSet to the board, changed move generation, added a bad search function. still running into lots of bugs. Nothing working yet.

+++ Gambitfish/engine/evaluate/evaluator.go
@@ -5,5 +5,5 @@ package evaluate
 import "../../game"
 
 type Evaluator interface {
-	Evaluate(*game.Board) float
+	Evaluate(*game.Board) float64
 }
